Refactoring + Some Editor Stuff

+++ Licht.Impl/Time/DefaultTimer.cs
@@ -9,22 +9,22 @@ namespace Licht.Impl.Time
     [PublicAPI]
     public class DefaultTimer : ITimer
     {
-        private readonly Func<double> _timeStepFn;
-        public DefaultTimer(Func<double> timeStepFn, int framesPerSecond = 60)
+        private readonly Func<float> _timeStepFn;
+        public DefaultTimer(Func<float> timeStepFn, int framesPerSecond = 60)
         {
-            _last = new Caterpillar<double> { TailSize = 1 };
+            _last = new Caterpillar<float> { TailSize = 1 };
             _timeStepFn = timeStepFn;
             FramesPerSecond = framesPerSecond;
         }
 
-        private double _offset;
-        private readonly Caterpillar<double> _last;
-        private double _elapsed;
+        private float _offset;
+        private readonly Caterpillar<float> _last;
+        private float _elapsed;
 
         public int FramesPerSecond { get; set; }
-        public double Multiplier { get; set; } = 1d;
-        public double TotalElapsedTimeInMilliseconds => _elapsed + _offset;
-        public double UpdatedTimeInMilliseconds => _last.Current;
+        public float Multiplier { get; set; } = 1f;
+        public float TotalElapsedTimeInMilliseconds => _elapsed + _offset;
+        public float UpdatedTimeInMilliseconds => _last.Current;
         public bool IsActive { get; private set; } = true;
         public bool Debug { get; set; }
 
@@ -44,16 +44,16 @@ namespace Licht.Impl.Time
         {
             if (!IsActive) return true;
 
-            _offset = 0d;
-            _elapsed = 0d;
+            _offset = 0f;
+            _elapsed = 0f;
 
             return true;
         }
 
-        public bool Set(double time)
+        public bool Set(float time)
         {
             _offset = time;
-            _elapsed = 0d;
+            _elapsed = 0f;
 
             return true;
         }
@@ -63,7 +63,7 @@ namespace Licht.Impl.Time
             if (!IsActive) return;
 
             var latest = _timeStepFn();
-            var frameDuration = 1d / FramesPerSecond * 1000d;
+            var frameDuration = 1f / FramesPerSecond * 1000f;
 
             var difference = frameDuration - latest;
