+++ Licht.Unity/Time/UnityTimer.cs
@@ -8,16 +8,16 @@ namespace Licht.Unity.Time
     {
         public UnityTimer()
         {
-            _last = new Caterpillar<double> { TailSize = 1 };
+            _last = new Caterpillar<float> { TailSize = 1 };
         }
 
-        private double _offset;
-        private readonly Caterpillar<double> _last;
-        private double _elapsed;
+        private float _offset;
+        private readonly Caterpillar<float> _last;
+        private float _elapsed;
 
-        public double Multiplier { get; set; } = 1d;
-        public double TotalElapsedTimeInMilliseconds => _elapsed + _offset;
-        public double UpdatedTimeInMilliseconds => _last.Current;
+        public float Multiplier { get; set; } = 1f;
+        public float TotalElapsedTimeInMilliseconds => _elapsed + _offset;
+        public float UpdatedTimeInMilliseconds => _last.Current;
         public bool IsActive { get; private set; } = true;
         public bool Debug { get; set; }
 
@@ -37,16 +37,16 @@ namespace Licht.Unity.Time
         {
             if (!IsActive) return true;
 
-            _offset = 0d;
-            _elapsed = 0d;
+            _offset = 0f;
+            _elapsed = 0f;
 
             return true;
         }
 
-        public bool Set(double time)
+        public bool Set(float time)
         {
             _offset = time;
-            _elapsed = 0d;
+            _elapsed = 0f;
 
             return true;
         }
@@ -55,7 +55,7 @@ namespace Licht.Unity.Time
         {
             if (!IsActive) return;
 
-            var latest = UnityEngine.Time.deltaTime * Multiplier * 1000d;
+            var latest = UnityEngine.Time.deltaTime * Multiplier * 1000f;
             _elapsed += latest;
             _last.Current = latest;
